- Extract BpmnMxGraph class in separate file - Replace all mxGraph type by BpmnMxGraph
